resource de posicao do bloco vazio adicionado
